backend/requirements.txt
Flask
firebase-admin
google-generativeai